Fix HTTP/HTTPS issue as USPS does not accept HTTPS as we know it. iOS9 requires HTTPS in ATS, so info.plist must specify HTTP allowed. Update tests.
