+++ PackageTracker/PackageTracker/ViewController.swift
@@ -19,7 +19,6 @@ class ViewController: UIViewController, UITableViewDataSource {
     override func viewDidLoad() {
         super.viewDidLoad()
         // Do any additional setup after loading the view, typically from a nib.
-        items = ["a", "b", "c"]
     }
